+++ packages/react/runtime/__test__/utils/globals.js
@@ -1,7 +1,7 @@
 // Copyright 2024 The Lynx Authors. All rights reserved.
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
-import { vi } from 'vitest';
+import { afterEach, beforeEach, expect, vi } from 'vitest';
 
 import { getJSModule } from './jsModule.ts';
 
@@ -35,6 +35,12 @@ const performance = {
   _bindPipelineIdWithTimingFlag: vi.fn((id, flag) => {
     performance.__functionCallHistory.push(['_bindPipelineIdWithTimingFlag', id, flag]);
   }),
+
+  profileStart: vi.fn(),
+  profileEnd: vi.fn(),
+  profileMark: vi.fn(),
+  profileFlowId: vi.fn(() => 666),
+  isProfileRecording: vi.fn(() => true),
 };
 
 class SelectorQuery {
@@ -132,5 +138,25 @@ function injectGlobals() {
   console.profileEnd = vi.fn();
   console.alog = vi.fn();
 }
+
+beforeEach(() => {
+  performance.profileStart.mockClear();
+  performance.profileEnd.mockClear();
+});
+
+afterEach((context) => {
+  const skippedTasks = [
+    // Skip preact/debug tests since it would throw errors and abort the rendering process
+    'preact/debug',
+    'should remove event listener when throw in cleanup',
+  ];
+  if (skippedTasks.some(task => context.task.name.includes(task))) {
+    return;
+  }
+
+  expect(performance.profileStart.mock.calls.length).toBe(
+    performance.profileEnd.mock.calls.length,
+  );
+});
 
 injectGlobals();