+++ packages/react/runtime/__test__/utils/jsModule.ts
@@ -14,14 +14,27 @@ class GlobalEventEmitter {
     }
     this.listeners[eventName] = this.listeners[eventName].filter((l) => l !== listener);
   }
-  emit(eventName: string, ...args: any[]): void {
+  emit(eventName: string, args: any[]): void {
     if (!this.listeners[eventName]) {
       return;
     }
-    this.listeners[eventName].forEach((listener) => listener(...args));
+    this.listeners[eventName].forEach((listener) => args ? listener(...args) : listener());
   }
   clear(): void {
     this.listeners = {};
+  }
+  removeAllListeners(eventName?: string): void {
+    if (eventName) {
+      delete this.listeners[eventName];
+    } else {
+      this.clear();
+    }
+  }
+  trigger(eventName: string, params: string | Record<any, any>): void {
+    this.emit(eventName, [params]);
+  }
+  toggle(eventName: string, ...data: unknown[]): void {
+    this.emit(eventName, data);
   }
 }