+++ packages/rspeedy/plugin-react/test/config.test.ts
@@ -3,17 +3,16 @@
 // LICENSE file in the root directory of this source tree.
 import path from 'node:path'
 
-import { createRsbuild } from '@rsbuild/core'
 import type { RsbuildInstance } from '@rsbuild/core'
 import { describe, expect, test, vi } from 'vitest'
 
 import type { ReactWebpackPlugin } from '@lynx-js/react-webpack-plugin'
-import { createRspeedy } from '@lynx-js/rspeedy'
 import type {
   LynxEncodePlugin,
   LynxTemplatePlugin,
 } from '@lynx-js/template-webpack-plugin'
 
+import { createStubRspeedy as createRspeedy } from './createRspeedy.js'
 import { pluginStubRspeedyAPI } from './stub-rspeedy-api.plugin.js'
 
 describe('Config', () => {
@@ -21,8 +20,8 @@ describe('Config', () => {
     vi.stubEnv('NODE_ENV', 'development')
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx(),
           pluginStubRspeedyAPI(),
@@ -185,8 +184,8 @@ describe('Config', () => {
     vi.stubEnv('NODE_ENV', 'production')
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx(),
           pluginStubRspeedyAPI(),
@@ -268,8 +267,8 @@ describe('Config', () => {
     vi.stubEnv('NODE_ENV', 'production')
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx(),
           pluginStubRspeedyAPI(),
@@ -293,8 +292,8 @@ describe('Config', () => {
     vi.stubEnv('NODE_ENV', 'production')
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx(),
           pluginStubRspeedyAPI(),
@@ -344,8 +343,8 @@ describe('Config', () => {
   test('enableRemoveCSSScope defaults to true', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx({}),
           pluginStubRspeedyAPI(),
@@ -368,8 +367,8 @@ describe('Config', () => {
   test('enableRemoveCSSScope can be set to undefined explicitly', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx({
             enableRemoveCSSScope: undefined,
@@ -415,8 +414,8 @@ describe('Config', () => {
   test('not sideEffects: false when enableRemoveCSSScope: false', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx({
             enableRemoveCSSScope: false,
@@ -441,8 +440,8 @@ describe('Config', () => {
   test('not sideEffects: false when enableRemoveCSSScope: true', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx({
             enableRemoveCSSScope: true,
@@ -467,8 +466,8 @@ describe('Config', () => {
   test('merged rsbuild config', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         plugins: [
           pluginReactLynx(),
           pluginStubRspeedyAPI(),
@@ -628,8 +627,8 @@ describe('Config', () => {
 
     test('legacy Rspeedy version (with `output.inlineScripts` defaults to `false`)', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -655,8 +654,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -698,8 +697,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               foo: './fixtures/basic.tsx',
@@ -756,8 +755,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               'foo/bar': './fixtures/basic.tsx',
@@ -814,8 +813,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               foo: './fixtures/basic.tsx',
@@ -892,8 +891,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -940,8 +939,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               foo: './fixtures/basic.tsx',
@@ -1019,8 +1018,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1065,8 +1064,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1112,8 +1111,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1158,8 +1157,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'production')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1204,8 +1203,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'development')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1251,8 +1250,8 @@ describe('Config', () => {
       vi.stubEnv('NODE_ENV', 'development')
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           source: {
             entry: {
               main: './fixtures/basic.tsx',
@@ -1302,8 +1301,8 @@ describe('Config', () => {
     test('default', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1319,8 +1318,8 @@ describe('Config', () => {
     test('with output.sourceMap: true', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           output: {
             sourceMap: true,
           },
@@ -1339,8 +1338,8 @@ describe('Config', () => {
     test('with output.sourceMap: false', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           output: {
             sourceMap: false,
           },
@@ -1359,8 +1358,8 @@ describe('Config', () => {
     test('with output.sourceMap.js: "nosources"', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           output: {
             sourceMap: {
               js: 'nosources-source-map',
@@ -1443,8 +1442,8 @@ describe('Config', () => {
     test('default', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1460,8 +1459,8 @@ describe('Config', () => {
     test('performance.chunkSplit.strategy: "all-in-one"', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1482,8 +1481,8 @@ describe('Config', () => {
     test('performance.chunkSplit.strategy: "split-by-size"', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1515,8 +1514,8 @@ describe('Config', () => {
     test('performance.chunkSplit.strategy: "split-by-experience"', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1560,8 +1559,8 @@ describe('Config', () => {
 
       let config, ReactLynxWebpackPlugin
 
-      let rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      let rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx({
               extractStr: true,
@@ -1588,8 +1587,8 @@ describe('Config', () => {
       expect(ReactLynxWebpackPlugin?.options.extractStr).toBe(false)
       // expect(config).toMatchInlineSnapshot()
 
-      rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx({
               extractStr: true,
@@ -1611,8 +1610,8 @@ describe('Config', () => {
       expect(ReactLynxWebpackPlugin?.options.extractStr).toBe(true)
       // expect(config).toMatchInlineSnapshot()
 
-      rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx({
               extractStr: true,
@@ -1643,8 +1642,8 @@ describe('Config', () => {
     test('tools.rspack.optimization.splitChunks: false', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1670,8 +1669,8 @@ describe('Config', () => {
     test('tools.rspack.optimization.splitChunks: {}', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1705,8 +1704,8 @@ describe('Config', () => {
     test('tools.rspack.optimization.splitChunks: {} without environment: "lynx"', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -1716,6 +1715,7 @@ describe('Config', () => {
               strategy: 'split-by-experience',
             },
           },
+          environments: {},
           tools: {
             rspack: {
               optimization: {
@@ -1737,8 +1737,8 @@ describe('Config', () => {
     test('tools.rspack.optimization.splitChunks: { chunks: "async" }', async () => {
       const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-      const rsbuild = await createRsbuild({
-        rsbuildConfig: {
+      const rsbuild = await createRspeedy({
+        rspeedyConfig: {
           plugins: [
             pluginReactLynx(),
             pluginStubRspeedyAPI(),
@@ -2143,8 +2143,8 @@ describe('MPA Config', () => {
   test('entries', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {
@@ -2207,8 +2207,8 @@ describe('MPA Config', () => {
   test('entry[]', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {
@@ -2273,8 +2273,8 @@ describe('MPA Config', () => {
   test('entry[] with environment.web', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { web: {} },
         source: {
           entry: {
@@ -2316,8 +2316,8 @@ describe('MPA Config', () => {
   test('entry.import', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {
@@ -2382,8 +2382,8 @@ describe('MPA Config', () => {
   test('entry.import[]', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {
@@ -2448,8 +2448,8 @@ describe('MPA Config', () => {
   test('entry.dependsOn', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {
@@ -2515,8 +2515,8 @@ describe('MPA Config', () => {
   test('entry.dependsOn[]', async () => {
     const { pluginReactLynx } = await import('../src/pluginReactLynx.js')
 
-    const rsbuild = await createRsbuild({
-      rsbuildConfig: {
+    const rsbuild = await createRspeedy({
+      rspeedyConfig: {
         environments: { lynx: {} },
         source: {
           entry: {