+++ packages/web-platform/web-core-server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @lynx-js/web-core-server
 
+## 0.15.5
+
+### Patch Changes
+
+- fix: load main-thread chunk in ESM format ([#1437](https://github.com/lynx-family/lynx-stack/pull/1437))
+
+  See [nodejs/node#59362](https://github.com/nodejs/node/issues/59362) for more details.
+
 ## 0.15.4
 
 ### Patch Changes