+++ packages/web-platform/web-mainthread-apis/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lynx-js/web-mainthread-apis
 
+## 0.15.5
+
+### Patch Changes
+
+- fix: load main-thread chunk in ESM format ([#1437](https://github.com/lynx-family/lynx-stack/pull/1437))
+
+  See [nodejs/node#59362](https://github.com/nodejs/node/issues/59362) for more details.
+
+- Updated dependencies [[`29434ae`](https://github.com/lynx-family/lynx-stack/commit/29434aec853f14242f521316429cf07a93b8c371), [`fb7096b`](https://github.com/lynx-family/lynx-stack/commit/fb7096bb3c79166cd619a407095b8206eccb7918)]:
+  - @lynx-js/web-constants@0.15.5
+  - @lynx-js/web-style-transformer@0.15.5
+
 ## 0.15.4
 
 ### Patch Changes